renamed file name to FarmShelterTests
