+++ src/test/java/FarmShelterTests/ChickenCoopTest.java
@@ -1,4 +1,4 @@
-package Farm;
+package FarmShelterTests;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
