+++ src/test/java/FarmShelterTests/StableTest.java
@@ -1,4 +1,4 @@
-package Farm;
+package FarmShelterTests;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
